\name{waldwolf-package}
\alias{waldwolf-package}
\alias{waldwolf}
\docType{package}
\title{
Package WaldWolf
}
\description{
Package destined for the sequence of the Wald-Wolfowitz test.
}
\details{
\tabular{ll}{
Package: \tab waldwolf\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-11-27\cr
License: \tab GPL\cr
}
The package includes a function walwolf.test
}
\author{
Jodavid Ferreira and Claudio Souza

Maintainer:  <contato@jodavidferreira.com>

}
\references{
Pedro A. Morettin e Clelia M. C. Toloi, 
\emph{Analise de
SeriesTemporais}, Segunda edicao, 2006.
}

\keyword{ package }